Time Comparison:
- Pass a config object instead of raw bollean flags to the time comparison control so the intention is clear when using it
- Split getTimeOffset tests so parseDttmToDate has its own file
- use inlined test instead of blocks with describe and it
