+++ superset-frontend/plugins/plugin-chart-echarts/src/BigNumber/BigNumberPeriodOverPeriod/controlPanel.ts
@@ -108,7 +108,11 @@ const config: ControlPanelConfig = {
         ],
       ],
     },
-    sections.timeComparisonControls(false, false, false),
+    sections.timeComparisonControls({
+      multi: false,
+      showCalculationType: false,
+      showFullChoices: false,
+    }),
   ],
   controlOverrides: {
     y_axis_format: {
